Init unit test

+++ Code/company/CTM/Customer/repository/VSoft.Company.CTM.Customer.Repository.Efc/Services/ICustomerRepositoryEfc.cs
@@ -5,7 +5,7 @@ using VSoft.Company.CTM.Customer.Repository.Services;
 
 namespace VSoft.Company.CTM.Customer.Repository.Efc.Services;
 
-public interface ICustomerRepositoryEfc : ICustomerRepository, IEfcRepositoryEntityMgmtId<CustomerDbContext, MCustomerEntity, int>
+public interface ICustomerRepositoryEfc : ICustomerRepository, IEfcRepositoryEntityMgmtId<CustomerDbContext, MCustomerEntity, long>
 {
    
 }
